Add tests and refactor code to support this
- As well, add heif2jpg_test.exe to install target; running it there
  means it will have the necessary .dlls it needs to not fail
